Add a big todo with big design implications.
That is, playback is controlled via REST per RJ and James'
implementation, but in this case since we're in AS3,
we can easily delegate to the plugin that provided the
result to handle the simplest/most complex cases
(for example, local playback, ping a service to get a singed URL,
have the plugin delegate even further to an RTMP container/SWFLoader,
Stratus chunk and transfer the payload, etc).  Neat.